Move page implementations into route modules
Inline the actual component code from pages/ into the route modules
in routes/ instead of re-exporting. Delete the old page files and
update test imports to reference the new locations. The pages/index.ts
barrel now re-exports from routes/ for backwards compatibility.
