+++ web/src/pages/__tests__/Profile.test.tsx
@@ -2,8 +2,8 @@ import { render } from "vitest-browser-react";
 import { page } from "vitest/browser";
 import { describe, test, expect, vi, beforeEach } from "vitest";
 import { createMemoryRouter, RouterProvider } from "react-router";
-import { Profile, profileAction } from "../Profile";
-import type { ProfilePostsResponse } from "../Profile";
+import { Profile, action as profileAction } from "@/routes/profile";
+import type { ProfilePostsResponse } from "@/routes/profile";
 import type { NewPost } from "@/types";
 import placeholderUrl from "./fixtures/placeholder.png";
